requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
opencv-python>=4.5.3
kaggle>=1.5.12
tqdm>=4.62.3
pillow>=8.3.1
pandas>=1.3.3 